Add development stage to Dockerfile for ECommerceNetApp
Updated Dockerfile to include a development stage for building and running the ECommerceNetApp.Api project. Created new directories for data and logs, built the project in Debug configuration, and set the working directory to the output directory. Copied appsettings files and added a command to generate a development certificate if it doesn't exist. Modified CMD to run the built DLL instead of using `dotnet run`.